+++ rktl_autonomy/CMakeLists.txt
@@ -13,5 +13,5 @@ catkin_package()
 # Tests
 if(CATKIN_ENABLE_TESTING)
   find_package(rostest REQUIRED)
-  add_rostest(test/test_rktl_rewards.launch)
+  add_rostest(test/test_step.test)
 endif()
